update renderer params and create unique dir for images

+++ utils/image_utils.py
@@ -9,7 +9,7 @@ def plot_image_observation(ax, img_mkd, size=None):
     if img_mkd.shape[2] == 1:  # plot an occupancy grid image
         ax.imshow(img_mkd[:, :, 0], cmap='gray', extent=(0, size, -size/2.0, size/2.0))
     elif img_mkd.shape[2] == 3:  # plot an rgb image
-        ax.imshow(img_mkd)
+        ax.imshow(img_mkd.astype(np.int32))
         ax.grid(False)
     else:
         raise NotImplementedError
